+++ src/Agents.Net.Tests/Tools/Communities/LegacyServiceBridgeCommunity/Messages/ServiceParameterPassed.cs
@@ -1,4 +1,9 @@
-﻿using System;
+﻿#region Copyright
+//  Copyright (c) Tobias Wilker and contributors
+//  This file is licensed under MIT
+#endregion
+
+using System;
 using System.Collections.Generic;
 using Agents.Net;