+++ src/Agents.Net/MessageGate.cs
@@ -6,6 +6,7 @@
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading;
 
 namespace Agents.Net
@@ -17,7 +18,7 @@ namespace Agents.Net
     /// <typeparam name="TEnd">Type of the end message.</typeparam>
     /// <remarks>
     /// <para>This helper should be used with caution as it breaks the basic concept of the agent framework,
-    /// that each agent does exactly one thing. There are only two use cases were this helper should be used.
+    /// that each agent does exactly one thing. There are only three use cases were this helper should be used.
     /// <list type="number">
     /// <item>
     ///     <term><b>An agent serving as a legacy bridge:</b><br/></term>
@@ -28,6 +29,14 @@ namespace Agents.Net
     ///         To achieve that the service agent needs to wait inside the service call for a specific end message.
     ///         In addition to that, the service call need to terminate once an exception message was send. It would
     ///         wait forever otherwise.
+    ///         Use <see cref="SendAndAwait"/> for that use case.
+    ///     </description>
+    /// </item>
+    /// <item>
+    ///     <term><b>When a single agent's task is to handle message pairs:</b><br/></term>
+    ///     <description>
+    ///         This is necessary for example when handling transactions. In this case the gate can be used to determine whether to rollback the transaction or not. Another example would be when an <see cref="InterceptorAgent"/> that delays the execution of a message so that an injected message chain can be executed.
+    ///         Use <see cref="SendAndContinue"/> for that use case.
     ///     </description>
     /// </item>
     /// <item>
@@ -56,6 +65,7 @@ namespace Agents.Net
     /// </code>
     ///         Similar are operations such as executing an external process or database operations are more examples
     ///         were the second use case would helpful.
+    ///         Use <see cref="SendAndContinue"/> for that use case.
     ///     </description>
     /// </item>
     /// </list>
@@ -152,7 +162,13 @@ namespace Agents.Net
         /// imitated service call is an async method. By default no CancellationToken will be used.
         /// </param>
         /// <returns>The <see cref="MessageGateResult{TEnd}"/> of the operation.</returns>
-        /// <remarks>For an example how to use this class see the type documentation.</remarks>
+        /// <remarks>
+        /// <para>For an example how to use this class see the type documentation.</para>
+        /// <para>
+        /// WARNING: Extensive use of this method will lead to time gaps in the execution. See .net issue on github: https://github.com/dotnet/runtime/issues/55562
+        /// Use this method only for the legacy service call. Of all other scenarios use <see cref="SendAndContinue"/>.
+        /// </para>
+        /// </remarks>
         public MessageGateResult<TEnd> SendAndAwait(TStart startMessage, Action<Message> onMessage, int timeout = NoTimout,
                                                     CancellationToken cancellationToken = default)
         {
@@ -164,6 +180,56 @@ namespace Agents.Net
             if (onMessage == null)
             {
                 throw new ArgumentNullException(nameof(onMessage));
+            }
+            
+            using (ManualResetEventSlim resetEvent = new ManualResetEventSlim())
+            {
+                ManualResetEventSlim local = resetEvent;
+                MessageGateResult<TEnd> result = null;
+                SendAndContinue(startMessage, onMessage, r =>
+                {
+                    result = r;
+                    local.Set();
+                },timeout, cancellationToken);
+                
+                //Do not use cancellation token here. The this is handled in the SendAndContinueMethod
+                resetEvent.Wait();
+                return result;
+            }
+        }
+
+        /// <summary>
+        /// The method to send a start message and wait for the end message.
+        /// </summary>
+        /// <param name="startMessage">The start message.</param>
+        /// <param name="onMessage">The action to send the message.</param>
+        /// <param name="continueAction">The action to execute once a <see cref="MessageGateResult{TEnd}"/> was created.</param>
+        /// <param name="timeout">
+        /// Optionally a timeout after which the method will return, without sending the result.
+        /// By default the timeout is <see cref="NoTimout"/>
+        /// </param>
+        /// <param name="cancellationToken">
+        /// Optionally a cancellation token to cancel the continue operation. By default no CancellationToken will be used.
+        /// </param>
+        /// <remarks>
+        /// <para>For an example how to use this class see the type documentation.</para>
+        /// </remarks>
+        public void SendAndContinue(TStart startMessage, Action<Message> onMessage, Action<MessageGateResult<TEnd>> continueAction, int timeout = NoTimout,
+                                    CancellationToken cancellationToken = default)
+        {
+            if (startMessage == null)
+            {
+                throw new ArgumentNullException(nameof(startMessage));
+            }
+
+            if (onMessage == null)
+            {
+                throw new ArgumentNullException(nameof(onMessage));
+            }
+
+            if (continueAction == null)
+            {
+                throw new ArgumentNullException(nameof(continueAction));
             }
 
             CancellationToken userCancelToken = cancellationToken;
@@ -185,24 +251,29 @@ namespace Agents.Net
             CancellationTokenSource combinedSource = CancellationTokenSource.CreateLinkedTokenSource(userCancelToken, timeoutCancelToken, exceptionCancelToken);
             cancellationToken = combinedSource.Token;
             sources.Add(combinedSource);
+            
             lock (exceptions)
             {
                 exceptions.Add(startMessage, new List<ExceptionMessage>());
             }
 
-            try
+            MessageDomain.CreateNewDomainsFor(startMessage);
+            exceptionCollector.Push(startMessage);
+            CancellationTokenRegistration register = default;
+            pairCollector.PushAndContinue(startMessage, set =>
             {
-                MessageDomain.CreateNewDomainsFor(startMessage);
-                exceptionCollector.Push(startMessage);
-                onMessage(startMessage);
+                set.MarkAsConsumed(set.Message1);
+                set.MarkAsConsumed(set.Message2);
+                TEnd endMessage = set.Message2;
+                    
+                MessageDomain.TerminateDomainsOf(startMessage);
                 
-                TEnd endMessage = null;
-                pairCollector.PushAndExecute(startMessage, set =>
-                {
-                    set.MarkAsConsumed(set.Message1);
-                    set.MarkAsConsumed(set.Message2);
-                    endMessage = set.Message2;
-                }, cancellationToken);
+                Dispose();
+                MessageGateResult<TEnd> result = new MessageGateResult<TEnd>(MessageGateResultKind.Success, endMessage, Enumerable.Empty<ExceptionMessage>());
+                continueAction(result);
+            }, cancellationToken);
+            register = cancellationToken.Register(() =>
+            {
                 MessageDomain.TerminateDomainsOf(startMessage);
                 MessageGateResultKind resultKind = MessageGateResultKind.Success;
                 if (userCancelToken.IsCancellationRequested)
@@ -217,19 +288,29 @@ namespace Agents.Net
                 {
                     resultKind = MessageGateResultKind.Timeout;
                 }
+
+                IEnumerable<ExceptionMessage> currentExceptions;
                 lock (exceptions)
                 {
-                    IEnumerable<ExceptionMessage> currentExceptions = exceptions[startMessage];
+                    currentExceptions = exceptions[startMessage];
                     exceptions.Remove(startMessage);
-                    return new MessageGateResult<TEnd>(resultKind, endMessage, currentExceptions);
                 }
-            }
-            finally
+
+                Dispose();
+                MessageGateResult<TEnd> result = new MessageGateResult<TEnd>(resultKind, null, currentExceptions);
+                continueAction(result);
+            });
+                
+            onMessage(startMessage);
+
+            void Dispose()
             {
                 foreach (CancellationTokenSource tokenSource in sources)
                 {
                     tokenSource.Dispose();
                 }
+                
+                register.Dispose();
 
                 lock (exceptionCancelTokens)
                 {