+++ CHANGELOG.md
@@ -17,6 +17,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Added
 - #102 - A new helper class `MessageGate` was introduced which helps covering mainly legacy uses cases
+- `InterceptionAction.Delay` - this can delay the execution of a message until a token is released. `InterceptionAction` is no longer an enum, but this does not break existing code.
+
+### Changed
+- `PushAndExecute` is now obsolete. It was replaced by the new `MessageGate` type. It will be removed with 2022.6.0
 
 ## 2021.6.2
 ### Added